Revert and added new method - Hover to see -
This should fix the creative mode bug. We also reverted the method that we did before as this new method we did could fix the Issue with hitting a player and you're in creative mode would mess up the flight that creative would usually have. There is no real fix of this even if you have the isCreative method when an event is disabling someone's flight when the damager is hitting the player, even in creative mode. It's because it recognises a player is in survival mode, and whatever mode it is, it would just think this game mode is like any other game mode in MCPE.
I'll be testing this to see if this is the fix or not.